增加用户update delete changePassword接口

+++ controllers/user.controller.go
@@ -22,22 +22,38 @@ func NewUserController(userSerivce services.UserService) UserController {
     }
 }
 
+/* 创建 */
 func (uc *UserController) CreateUser(ctx *gin.Context) {
     var user models.User
     if err := ctx.ShouldBindJSON(&user); err != nil {
-        ctx.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
+        response := gin.H{
+            "code":    10000,
+            "message": "请求参数异常",
+            "error":   err.Error(),
+        }
+        ctx.JSON(http.StatusBadGateway, response)
         return
     }
 
     err := uc.UserService.CreateUser(&user)
     if err != nil {
-        ctx.JSON(http.StatusBadGateway, gin.H{"message": err.Error()})
+        response := gin.H{
+            "code":    10000,
+            "message": "服务处理异常",
+            "error":   err.Error(),
+        }
+        ctx.JSON(http.StatusBadGateway, response)
         return
     }
 
-    ctx.JSON(http.StatusOK, gin.H{"message": "success"})
+    response := gin.H{
+        "code":    20000,
+        "message": "用户成功创建",
+    }
+    ctx.JSON(http.StatusOK, response)
 }
 
+/* 获取 */
 func (uc *UserController) GetUser(ctx *gin.Context) {
     var id string
     id = ctx.Query("id")
@@ -77,34 +93,68 @@ func (uc *UserController) GetUser(ctx *gin.Context) {
     ctx.JSON(http.StatusOK, response)
 }
 
+/* 更新 */
 func (uc *UserController) UpdateUser(ctx *gin.Context) {
     var user models.User
     if err := ctx.ShouldBindJSON(&user); err != nil {
-        ctx.JSON(http.StatusBadRequest, gin.H{"message": err.Error()})
+        response := gin.H{
+            "code":    10000,
+            "message": "请求参数异常",
+            "error":   err.Error(),
+        }
+        ctx.JSON(http.StatusBadGateway, response)
         return
     }
 
     err := uc.UserService.UpdateUser(&user)
     if err != nil {
-        ctx.JSON(http.StatusBadGateway, gin.H{"message": err.Error()})
+        response := gin.H{
+            "code":    10000,
+            "message": "服务处理异常",
+            "error":   err.Error(),
+        }
+        ctx.JSON(http.StatusBadGateway, response)
         return
     }
 
-    ctx.JSON(http.StatusOK, gin.H{"message": "success"})
+    response := gin.H{
+        "code":    20000,
+        "message": "用户成功更新",
+    }
+    ctx.JSON(http.StatusOK, response)
 }
 
+/* 删除 */
 func (uc *UserController) DeleteUser(ctx *gin.Context) {
     id := ctx.Query("id")
+    if len(id) == 0 {
+        response := gin.H{
+            "code":    10000,
+            "message": "请求参数异常",
+        }
+        ctx.JSON(http.StatusBadGateway, response)
+        return
+    }
 
     err := uc.UserService.DeleteUser(&id)
     if err != nil {
-        ctx.JSON(http.StatusBadGateway, gin.H{"message": err.Error()})
+        response := gin.H{
+            "code":    10000,
+            "message": "服务处理异常",
+            "error":   err.Error(),
+        }
+        ctx.JSON(http.StatusBadGateway, response)
         return
     }
 
-    ctx.JSON(http.StatusOK, gin.H{"message": "success"})
+    response := gin.H{
+        "code":    20000,
+        "message": "用户成功删除",
+    }
+    ctx.JSON(http.StatusOK, response)
 }
 
+/* 获取列表 */
 func (uc *UserController) GetUserList(ctx *gin.Context) {
     var page, limit, sort, status, keyword string
     page = ctx.DefaultQuery("page", "0")
@@ -151,6 +201,37 @@ func (uc *UserController) GetUserList(ctx *gin.Context) {
         "code":    20000,
         "message": "用户列表成功获取",
         "data":    data,
+    }
+    ctx.JSON(http.StatusOK, response)
+}
+
+/* 更改用户密码 */
+func (uc *UserController) ChangeUserPassword(ctx *gin.Context) {
+    var user models.UserPassword
+    if err := ctx.ShouldBindJSON(&user); err != nil {
+        response := gin.H{
+            "code":    10000,
+            "message": "请求参数异常",
+            "error":   err.Error(),
+        }
+        ctx.JSON(http.StatusBadGateway, response)
+        return
+    }
+
+    err := uc.UserService.ChangeUserPassword(&user)
+    if err != nil {
+        response := gin.H{
+            "code":    10000,
+            "message": "服务处理异常",
+            "error":   err.Error(),
+        }
+        ctx.JSON(http.StatusBadGateway, response)
+        return
+    }
+
+    response := gin.H{
+        "code":    20000,
+        "message": "用户密码成功更新",
     }
     ctx.JSON(http.StatusOK, response)
 }
